+++ spec/lib/kapnismology/evaluation_spec.rb
@@ -4,7 +4,7 @@ module Kapnismology
   RSpec.describe Evaluation do
     class TestSmokeTest
       def result
-        Result.new(true, [:berserk], '黄金時代') # defined down there in the let
+        Result.new(true, { title: 'berserk' }, '黄金時代') # defined down there in the let
       end
 
       def __result__
@@ -14,7 +14,7 @@ module Kapnismology
     let(:evaluation) { Evaluation.new(TestSmokeTest) }
 
     it 'creates a json representation' do
-      expected = '{"TestSmokeTest":{"passed":true,"data":["berserk"],"message":"黄金時代","extra_messages":[]}}'
+      expected = '{"name":"test_smoke_test","passed":true,"data":{"title":"berserk"},"message":"黄金時代","debug_messages":[]}'
       expect(evaluation.to_json).to eq(expected)
     end
 
@@ -23,7 +23,7 @@ module Kapnismology
     end
 
     it 'creates a string representation' do
-      expected = "#{Terminal.green('passed')}: TestSmokeTest\n\e[1m黄金時代\e[0m\n   [:berserk]\n"
+      expected = %|#{Terminal.green('passed')}: TestSmokeTest\n\e[1m黄金時代\e[0m\n   {:title=>"berserk"}\n|
       expect(evaluation.to_s).to eq(expected)
     end
   end